fix(RV-497): Fix 1st time ux to templates > 0
