+++ frontend/src/components/ExtractUploadFile.scss
@@ -57,7 +57,7 @@ p {
   display: flex;
   flex-direction: column;
   height: 150px;
-  width: 914px;
+  width: 100%;
   border-left: 5px solid #d63e04;
   background-color: #f4e3db;
 }
